✨ Add new filter configurations
